docs: add troubleshooting for podman user namespaces
Adds a new troubleshooting section to the README to address a common issue when using Podman to build images within a container. The error is related to user namespaces and can be resolved by configuring /etc/subuid and /etc/subgid.

This section provides a clear explanation and a Dockerfile snippet to fix the problem.

Signed-off-by: Zeglius <33781398+Zeglius@users.noreply.github.com>